Implemented traffic light system for first two field sets
