+++ plugin/napari_lattice/icons.py
@@ -1,0 +1,5 @@
+import pkg_resources
+
+GREEN = pkg_resources.resource_filename(__name__, "green.png")
+GREY = pkg_resources.resource_filename(__name__, "grey.png")
+RED = pkg_resources.resource_filename(__name__, "red.png")
